refactor(fib-js): add bridge and fleet
