+++ bin/.env.example
@@ -1,0 +1,7 @@
+OP_HTTP_ENDPOINT=http://127.0.0.1:8545
+OP_WS_ENDPOINT=ws://127.0.0.1:8545
+BASE_ENDPOINT=http://127.0.0.1:8545
+MANAGER_CONTRACT_ADDRESS=0xCf7Ed3AccA5a467e9e704C703E8D87F634fB0Fc9
+OUTPOST_CONTRACT_ADDRESS=0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512
+LOG_START_BLOCK=1
+ER_BASE_ENDPOINT=http://host.docker.internal:8545
