+++ fibonacci-js/contract/.env.example
@@ -1,3 +1,33 @@
-APP_CONTRACT=0x5FbDB2315678afecb367f032d93F642f64180aa3
+# for testing
+# fleet address to settle the result to contract
+FLEET_SETTLER_ADDRESS=0x3fCF1a0A658Cbd89577DFb2559a99A68bd2B8c80
+
+
+# test app docker image, build on local first
+DOCKER_URI=abci-fib:latest
+DOCKER_HASH=sha256:d10e26f90061400174e3f48f16da183c5ea2e2595c1fb050edfae1cdd4322f21
+
+
+# RPC
+OPTIMISM_RPC=http://127.0.0.1:8545
+BASE_RPC=http://127.0.0.1:8545
+
+# addresses for deployment, can't be the same address
+## used for proxy admin
+PROXY_ADMIN_PRIVATE_KEY=0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80
+PROXY_ADMIN_ADDRESS=0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266
+## used for implementation contract owner
+CONTRACT_OWNER_PRIVATE_KEY=0x59c6995e998f97a5a0044966f0945389dc9e86dae88c7a8412f4603b6b78690d
+CONTRACT_OWNER_ADDRESS=0x70997970C51812dc3A010C7d01b50e0d17dc79C8
+
+# hardhat local testnet contracts
+## outpost
 OUTPOST_PROXY_CONTRACT=0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512
-DEPLOYER_PRIVATE_KEY=0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80
+OUTPOST_PROXY_ADMIN_CONTRACT=0xCafac3dD18aC6c6e92c921884f9E4176737C052c
+
+## manager
+MANAGER_PROXY_CONTRACT=0xCf7Ed3AccA5a467e9e704C703E8D87F634fB0Fc9
+MANAGER_PROXY_ADMIN_CONTRACT=0xd8058efe0198ae9dD7D563e1b4938Dcbc86A1F81
+
+## app
+APP_CONTRACT=0xDc64a140Aa3E981100a9becA4E685f962f0cF6C9
