+++ fibonacci-js/contract/APPTest.md
@@ -1,0 +1,12 @@
+
+## request a fibNum
+```
+// request 10
+make request-fib NUM=10
+```
+
+## check the result
+Wait for the bridge and fleet settle the result back
+```
+make fib-result NUM=10
+```
