--- fibonacci-js/contract/ignition/modules/MockOutpost.ts
@@ -1,10 +1,0 @@
-import { buildModule } from "@nomicfoundation/hardhat-ignition/modules";
-
-
-const mockOutpostModule = buildModule("MockOutpostModule", (m) => {
-    const outpost = m.contract("MockOutpost");
-
-    return { outpost };
-});
-
-export default mockOutpostModule;
