+++ fibonacci-js/server/README.md
@@ -1,17 +1,7 @@
 # Local Testing  
 
-`INIT_DATA` represents the ABI-encoded data from the contract.  
-
-For example, if we want to compute `fib(22)`, the corresponding `INIT_DATA` should be:  
-
-```sh
-0x0000000000000000000000000000000000000000000000000000000000000016
+## build docker image
 ```
-
-You can retrieve the **testing `INIT_DATA`** from contract testing and run the following command:  
-
-```sh
-INIT_DATA=0x0000000000000000000000000000000000000000000000000000000000000016 make run
+docker build -t abci-fib:latest .
 ```
-
-The **encoded result** obtained from this process is the data that will be settled back into the contract.  
+Wait for the image building done.
